+++ woodwork/tests/data_table/test_datatable.py
@@ -46,7 +46,7 @@ from woodwork.tests.testing_utils import (
 
 def test_datatable_init(sample_df):
     dt = DataTable(sample_df)
-    df = dt.to_pandas()
+    df = dt.to_dataframe()
 
     assert dt.name is None
     assert dt.index is None
@@ -59,10 +59,10 @@ def test_datatable_init(sample_df):
 
 def test_datatable_copy_param(sample_df):
     dt_with_copy = DataTable(sample_df, copy_dataframe=True)
-    assert sample_df is not dt_with_copy.to_pandas()
+    assert sample_df is not dt_with_copy.to_dataframe()
 
     dt_no_copy = DataTable(sample_df)
-    assert sample_df is dt_no_copy.to_pandas()
+    assert sample_df is dt_no_copy.to_dataframe()
 
 
 def test_datatable_init_with_name_and_index_vals(sample_df):
@@ -97,7 +97,7 @@ def test_datatable_with_numeric_datetime_time_index(time_index_df):
         DataTable(time_index_df, name='datatable', time_index='strs', logical_types={'strs': Datetime})
 
     assert dt.time_index == 'ints'
-    assert dt.to_pandas()['ints'].dtype == 'datetime64[ns]'
+    assert dt.to_dataframe()['ints'].dtype == 'datetime64[ns]'
 
 
 def test_datatable_with_numeric_time_index(time_index_df):
@@ -660,7 +660,7 @@ def test_sets_category_dtype_on_init():
             dt = DataTable(pd.DataFrame(series), logical_types=ltypes)
             assert dt.columns[column_name].logical_type == logical_type
             assert dt.columns[column_name].dtype == logical_type.pandas_dtype
-            assert dt.to_pandas()[column_name].dtype == logical_type.pandas_dtype
+            assert dt.to_dataframe()[column_name].dtype == logical_type.pandas_dtype
 
 
 def test_sets_category_dtype_on_update():
@@ -683,7 +683,7 @@ def test_sets_category_dtype_on_update():
         dt = dt.set_logical_types({column_name: logical_type})
         assert dt.columns[column_name].logical_type == logical_type
         assert dt.columns[column_name].dtype == logical_type.pandas_dtype
-        assert dt.to_pandas()[column_name].dtype == logical_type.pandas_dtype
+        assert dt.to_dataframe()[column_name].dtype == logical_type.pandas_dtype
 
 
 def test_sets_string_dtype_on_init():
@@ -714,7 +714,7 @@ def test_sets_string_dtype_on_init():
             dt = DataTable(pd.DataFrame(series), logical_types=ltypes)
             assert dt.columns[column_name].logical_type == logical_type
             assert dt.columns[column_name].dtype == logical_type.pandas_dtype
-            assert dt.to_pandas()[column_name].dtype == logical_type.pandas_dtype
+            assert dt.to_dataframe()[column_name].dtype == logical_type.pandas_dtype
 
 
 def test_sets_string_dtype_on_update():
@@ -739,7 +739,7 @@ def test_sets_string_dtype_on_update():
         dt = dt.set_logical_types({column_name: logical_type})
         assert dt.columns[column_name].logical_type == logical_type
         assert dt.columns[column_name].dtype == logical_type.pandas_dtype
-        assert dt.to_pandas()[column_name].dtype == logical_type.pandas_dtype
+        assert dt.to_dataframe()[column_name].dtype == logical_type.pandas_dtype
 
 
 def test_sets_boolean_dtype_on_init():
@@ -760,7 +760,7 @@ def test_sets_boolean_dtype_on_init():
         dt = DataTable(pd.DataFrame(series), logical_types=ltypes)
         assert dt.columns[column_name].logical_type == logical_type
         assert dt.columns[column_name].dtype == logical_type.pandas_dtype
-        assert dt.to_pandas()[column_name].dtype == logical_type.pandas_dtype
+        assert dt.to_dataframe()[column_name].dtype == logical_type.pandas_dtype
 
 
 def test_sets_boolean_dtype_on_update():
@@ -774,7 +774,7 @@ def test_sets_boolean_dtype_on_update():
     dt = dt.set_logical_types({column_name: Boolean})
     assert dt.columns[column_name].logical_type == Boolean
     assert dt.columns[column_name].dtype == Boolean.pandas_dtype
-    assert dt.to_pandas()[column_name].dtype == Boolean.pandas_dtype
+    assert dt.to_dataframe()[column_name].dtype == Boolean.pandas_dtype
 
 
 def test_sets_int64_dtype_on_init():
@@ -796,7 +796,7 @@ def test_sets_int64_dtype_on_init():
             dt = DataTable(pd.DataFrame(series), logical_types=ltypes)
         assert dt.columns[column_name].logical_type == logical_type
         assert dt.columns[column_name].dtype == logical_type.pandas_dtype
-        assert dt.to_pandas()[column_name].dtype == logical_type.pandas_dtype
+        assert dt.to_dataframe()[column_name].dtype == logical_type.pandas_dtype
 
 
 def test_sets_int64_dtype_on_update():
@@ -816,7 +816,7 @@ def test_sets_int64_dtype_on_update():
         dt = dt.set_logical_types({column_name: logical_type})
         assert dt.columns[column_name].logical_type == logical_type
         assert dt.columns[column_name].dtype == logical_type.pandas_dtype
-        assert dt.to_pandas()[column_name].dtype == logical_type.pandas_dtype
+        assert dt.to_dataframe()[column_name].dtype == logical_type.pandas_dtype
 
 
 def test_sets_float64_dtype_on_init():
@@ -836,7 +836,7 @@ def test_sets_float64_dtype_on_init():
         dt = DataTable(pd.DataFrame(series), logical_types=ltypes)
         assert dt.columns[column_name].logical_type == logical_type
         assert dt.columns[column_name].dtype == logical_type.pandas_dtype
-        assert dt.to_pandas()[column_name].dtype == logical_type.pandas_dtype
+        assert dt.to_dataframe()[column_name].dtype == logical_type.pandas_dtype
 
 
 def test_sets_float64_dtype_on_update():
@@ -850,7 +850,7 @@ def test_sets_float64_dtype_on_update():
     dt = dt.set_logical_types({column_name: Double})
     assert dt.columns[column_name].logical_type == Double
     assert dt.columns[column_name].dtype == Double.pandas_dtype
-    assert dt.to_pandas()[column_name].dtype == Double.pandas_dtype
+    assert dt.to_dataframe()[column_name].dtype == Double.pandas_dtype
 
 
 def test_sets_datetime64_dtype_on_init():
@@ -872,7 +872,7 @@ def test_sets_datetime64_dtype_on_init():
         dt = DataTable(pd.DataFrame(series), logical_types=ltypes)
         assert dt.columns[column_name].logical_type == logical_type
         assert dt.columns[column_name].dtype == logical_type.pandas_dtype
-        assert dt.to_pandas()[column_name].dtype == logical_type.pandas_dtype
+        assert dt.to_dataframe()[column_name].dtype == logical_type.pandas_dtype
 
 
 def test_sets_datetime_dtype_on_update():
@@ -886,7 +886,7 @@ def test_sets_datetime_dtype_on_update():
     dt = dt.set_logical_types({column_name: Datetime})
     assert dt.columns[column_name].logical_type == Datetime
     assert dt.columns[column_name].dtype == Datetime.pandas_dtype
-    assert dt.to_pandas()[column_name].dtype == Datetime.pandas_dtype
+    assert dt.to_dataframe()[column_name].dtype == Datetime.pandas_dtype
 
 
 def test_invalid_dtype_casting():
@@ -930,7 +930,7 @@ def test_int_dtype_inference_on_init():
         'ints_nan': pd.Series([1, np.nan]),
         'ints_NA': pd.Series([1, pd.NA]),
         'ints_NA_specified': pd.Series([1, pd.NA], dtype='Int64')})
-    df_from_dt = DataTable(df).to_pandas()
+    df_from_dt = DataTable(df).to_dataframe()
 
     assert df_from_dt['ints_no_nans'].dtype == 'Int64'
     assert df_from_dt['ints_nan'].dtype == 'float64'
@@ -944,7 +944,7 @@ def test_bool_dtype_inference_on_init():
         'bool_nan': pd.Series([True, np.nan]),
         'bool_NA': pd.Series([True, pd.NA]),
         'bool_NA_specified': pd.Series([True, pd.NA], dtype="boolean")})
-    df_from_dt = DataTable(df).to_pandas()
+    df_from_dt = DataTable(df).to_dataframe()
 
     assert df_from_dt['bools_no_nans'].dtype == 'boolean'
     assert df_from_dt['bool_nan'].dtype == 'category'
@@ -961,7 +961,7 @@ def test_str_dtype_inference_on_init():
         'long_str_NA_specified': pd.Series(['this is a very long sentence inferred as a string', pd.NA], dtype="string"),
         'long_str_NA': pd.Series(['this is a very long sentence inferred as a string', pd.NA])
     })
-    df_from_dt = DataTable(df).to_pandas()
+    df_from_dt = DataTable(df).to_dataframe()
 
     assert df_from_dt['str_no_nans'].dtype == 'category'
     assert df_from_dt['str_nan'].dtype == 'category'
@@ -977,7 +977,7 @@ def test_float_dtype_inference_on_init():
         'floats_nan': pd.Series([1.1, np.nan]),
         'floats_NA': pd.Series([1.1, pd.NA]),
         'floats_nan_specified': pd.Series([1.1, np.nan], dtype='float')})
-    df_from_dt = DataTable(df).to_pandas()
+    df_from_dt = DataTable(df).to_dataframe()
 
     assert df_from_dt['floats_no_nans'].dtype == 'float64'
     assert df_from_dt['floats_nan'].dtype == 'float64'
@@ -992,7 +992,7 @@ def test_datetime_dtype_inference_on_init():
         'date_NA': pd.Series([pd.to_datetime('2020-09-01'), pd.NA]),
         'date_NaT': pd.Series([pd.to_datetime('2020-09-01'), pd.NaT]),
         'date_NA_specified': pd.Series([pd.to_datetime('2020-09-01'), pd.NA], dtype='datetime64[ns]')})
-    df_from_dt = DataTable(df).to_pandas()
+    df_from_dt = DataTable(df).to_dataframe()
 
     assert df_from_dt['date_no_nans'].dtype == 'datetime64[ns]'
     assert df_from_dt['date_nan'].dtype == 'datetime64[ns]'
@@ -1008,7 +1008,7 @@ def test_timedelta_dtype_inference_on_init():
         'delta_NaT': (pd.Series([pd.to_datetime('2020-09-01'), pd.NaT]) - pd.to_datetime('2020-07-01')),
         'delta_NA_specified': (pd.Series([pd.to_datetime('2020-09-01'), pd.NA], dtype='datetime64[ns]') - pd.to_datetime('2020-07-01')),
     })
-    df_from_dt = DataTable(df).to_pandas()
+    df_from_dt = DataTable(df).to_dataframe()
 
     assert df_from_dt['delta_no_nans'].dtype == 'timedelta64[ns]'
     assert df_from_dt['delta_nan'].dtype == 'timedelta64[ns]'
@@ -1041,7 +1041,7 @@ def test_select_ltypes_warning(sample_df):
     with pytest.warns(UserWarning, match=warning):
         dt_all_types = dt.select(all_types)
     assert len(dt_all_types.columns) == len(dt.columns)
-    assert len(dt_all_types.to_pandas().columns) == len(dt.to_pandas().columns)
+    assert len(dt_all_types.to_dataframe().columns) == len(dt.to_dataframe().columns)
 
 
 def test_select_ltypes_strings(sample_df):
@@ -1237,7 +1237,7 @@ def test_pop(sample_df):
     assert to_pandas(datacol.to_series()).values == [33, 25, 33, 57]
     assert datacol.logical_type == WholeNumber
 
-    assert 'age' not in dt.to_pandas().columns
+    assert 'age' not in dt.to_dataframe().columns
     assert 'age' not in dt.columns
 
     assert 'age' not in dt.logical_types.keys()
@@ -1291,13 +1291,13 @@ def test_getitem_invalid_input(sample_df):
 def test_datatable_getitem_list_input(sample_df):
     # Test regular columns
     dt = DataTable(sample_df, time_index='signup_date', index='id', name='dt_name')
-    df = dt.to_pandas()
+    df = dt.to_dataframe()
     columns = ['age', 'full_name']
     new_dt = dt[columns]
     assert new_dt is not dt
-    assert new_dt.to_pandas() is not df
-    pd.testing.assert_frame_equal(to_pandas(df[columns]).reset_index(drop=True), to_pandas(new_dt.to_pandas()))
-    assert all(new_dt.to_pandas().columns == ['age', 'full_name'])
+    assert new_dt.to_dataframe() is not df
+    pd.testing.assert_frame_equal(to_pandas(df[columns]).reset_index(drop=True), to_pandas(new_dt.to_dataframe()))
+    assert all(new_dt.to_dataframe().columns == ['age', 'full_name'])
     assert set(new_dt.columns.keys()) == {'age', 'full_name'}
     assert new_dt.index is None
     assert new_dt.time_index is None
@@ -1306,9 +1306,9 @@ def test_datatable_getitem_list_input(sample_df):
     columns = ['id', 'full_name']
     new_dt = dt[columns]
     assert new_dt is not dt
-    assert new_dt.to_pandas() is not df
-    pd.testing.assert_frame_equal(to_pandas(df[columns]), to_pandas(new_dt.to_pandas()))
-    assert all(new_dt.to_pandas().columns == ['id', 'full_name'])
+    assert new_dt.to_dataframe() is not df
+    pd.testing.assert_frame_equal(to_pandas(df[columns]), to_pandas(new_dt.to_dataframe()))
+    assert all(new_dt.to_dataframe().columns == ['id', 'full_name'])
     assert set(new_dt.columns.keys()) == {'id', 'full_name'}
     assert new_dt.index == 'id'
     assert new_dt.time_index is None
@@ -1317,16 +1317,18 @@ def test_datatable_getitem_list_input(sample_df):
     columns = ['id', 'signup_date', 'full_name']
     new_dt = dt[columns]
     assert new_dt is not dt
-    assert new_dt.to_pandas() is not df
-    pd.testing.assert_frame_equal(to_pandas(df[columns]), to_pandas(new_dt.to_pandas()), check_index_type=False)
+    assert new_dt.to_dataframe() is not df
+    pd.testing.assert_frame_equal(to_pandas(df[columns]), to_pandas(new_dt.to_dataframe()), check_index_type=False)
+    assert all(new_dt.to_dataframe().columns == ['id', 'signup_date', 'full_name'])
+    assert set(new_dt.columns.keys()) == {'id', 'signup_date', 'full_name'}
     assert new_dt.index == 'id'
 
     # Test with empty list selector
     columns = []
     new_dt = dt[columns]
     assert new_dt is not dt
-    assert new_dt.to_pandas() is not df
-    assert to_pandas(new_dt.to_pandas()).empty
+    assert new_dt.to_dataframe() is not df
+    assert to_pandas(new_dt.to_dataframe()).empty
     assert set(new_dt.columns.keys()) == set()
     assert new_dt.index is None
     assert new_dt.time_index is None
@@ -1378,7 +1380,7 @@ def test_setitem_different_name(sample_df):
 
     assert dt['id'].name == 'id'
     assert dt['id'].to_series().name == 'id'
-    assert dt.to_pandas()['id'].name == 'id'
+    assert dt.to_dataframe()['id'].name == 'id'
     assert 'wrong' not in dt.columns
 
     warning = 'Key, new_col, does not match the name of the provided DataColumn, wrong2.'\
@@ -1388,7 +1390,7 @@ def test_setitem_different_name(sample_df):
                                    use_standard_tags=False)
     assert dt['new_col'].name == 'new_col'
     assert dt['new_col'].to_series().name == 'new_col'
-    assert dt.to_pandas()['new_col'].name == 'new_col'
+    assert dt.to_dataframe()['new_col'].name == 'new_col'
     assert 'wrong2' not in dt.columns
 
 
@@ -1398,7 +1400,7 @@ def test_setitem_new_column(sample_df):
     new_col = DataColumn(pd.Series([1, 2, 3], dtype='Int64'),
                          use_standard_tags=False)
     dt['test_col2'] = new_col
-    updated_df = dt.to_pandas()
+    updated_df = dt.to_dataframe()
     assert 'test_col2' in dt.columns
     assert dt['test_col2'].logical_type == WholeNumber
     assert dt['test_col2'].semantic_tags == set()
@@ -1409,7 +1411,7 @@ def test_setitem_new_column(sample_df):
     new_col = DataColumn(pd.Series(['new', 'column', 'inserted'], dtype='string'),
                          use_standard_tags=True)
     dt['test_col'] = new_col
-    updated_df = dt.to_pandas()
+    updated_df = dt.to_dataframe()
     assert 'test_col' in dt.columns
     assert dt['test_col'].logical_type == Categorical
     assert dt['test_col'].semantic_tags == {'category'}
@@ -1422,7 +1424,7 @@ def test_setitem_new_column(sample_df):
                          use_standard_tags=False,
                          semantic_tags={'test_tag'})
     dt['test_col3'] = new_col
-    updated_df = dt.to_pandas()
+    updated_df = dt.to_dataframe()
     assert 'test_col3' in dt.columns
     assert dt['test_col3'].logical_type == Double
     assert dt['test_col3'].semantic_tags == {'test_tag'}
@@ -1440,7 +1442,7 @@ def test_setitem_overwrite_column(sample_df):
     overwrite_col = DataColumn(pd.Series([1, 2, 3], dtype='Int64'),
                                use_standard_tags=True)
     dt['age'] = overwrite_col
-    updated_df = dt.to_pandas()
+    updated_df = dt.to_dataframe()
 
     assert 'age' in dt.columns
     assert dt['age'].logical_type == original_col.logical_type
@@ -1455,7 +1457,7 @@ def test_setitem_overwrite_column(sample_df):
                                use_standard_tags=False,
                                semantic_tags='test_tag')
     dt['full_name'] = overwrite_col
-    updated_df = dt.to_pandas()
+    updated_df = dt.to_dataframe()
 
     assert 'full_name' in dt.columns
     assert dt['full_name'].logical_type == Boolean
@@ -1499,9 +1501,9 @@ def test_set_index(sample_df):
     if isinstance(sample_df, pd.DataFrame):
         dt = DataTable(sample_df)
         dt.index = 'id'
-        assert (dt.to_pandas().index == [0, 1, 2, 3]).all()
+        assert (dt.to_dataframe().index == [0, 1, 2, 3]).all()
         dt.index = 'full_name'
-        assert (dt.to_pandas().index == dt.to_pandas()['full_name']).all()
+        assert (dt.to_dataframe().index == dt.to_dataframe()['full_name']).all()
 
 
 def test_set_time_index(sample_df):
@@ -1795,11 +1797,11 @@ def test_natural_language_inference_with_config_options():
     ww.config.reset_option('natural_language_threshold')
 
 
-def test_to_pandas_copy(sample_df):
+def test_to_dataframe_copy(sample_df):
     dt = DataTable(sample_df)
 
-    df_no_copy = dt.to_pandas()
-    df_copy = dt.to_pandas(copy=True)
+    df_no_copy = dt.to_dataframe()
+    df_copy = dt.to_dataframe(copy=True)
 
     assert df_no_copy is sample_df
 
@@ -2103,7 +2105,7 @@ def test_data_table_handle_nans_for_mutual_info():
         'str_no_nan': pd.Series(['test', 'test2', 'test2', 'test']),
     })
     dt_nans = DataTable(df_nans)
-    formatted_df = dt_nans._handle_nans_for_mutual_info(dt_nans.to_pandas(copy=True))
+    formatted_df = dt_nans._handle_nans_for_mutual_info(dt_nans.to_dataframe(copy=True))
 
     assert isinstance(formatted_df, pd.DataFrame)
 
@@ -2124,7 +2126,7 @@ def test_data_table_make_categorical_for_mutual_info():
         'categories': pd.Series(['test', 'test2', 'test2', 'test'])
     })
     dt = DataTable(df)
-    formatted_num_bins_df = dt._make_categorical_for_mutual_info(dt.to_pandas(copy=True), num_bins=4)
+    formatted_num_bins_df = dt._make_categorical_for_mutual_info(dt.to_dataframe(copy=True), num_bins=4)
 
     assert isinstance(formatted_num_bins_df, pd.DataFrame)
 
@@ -2147,8 +2149,7 @@ def test_data_table_get_mutual_information(df_same_mi, df_mi):
     assert mi_between_cols('floats', 'ints', mi) == 1.0
 
     dt = DataTable(df_mi)
-    original_df = dt.to_pandas(copy=True)
-
+    original_df = dt.to_dataframe(copy=True)
     mi = dt.get_mutual_information()
     assert mi.shape[0] == 6
     np.testing.assert_almost_equal(mi_between_cols('ints', 'bools', mi), 0.734, 3)
@@ -2169,7 +2170,7 @@ def test_data_table_get_mutual_information(df_same_mi, df_mi):
     np.testing.assert_almost_equal(mi_between_cols('bools', 'strs', mi), 0, 3)
 
     # Confirm that none of this changed the DataTable's underlying df
-    pd.testing.assert_frame_equal(to_pandas(dt.to_pandas()), to_pandas(original_df))
+    pd.testing.assert_frame_equal(to_pandas(dt.to_dataframe()), to_pandas(original_df))
 
 
 def test_mutual_info_does_not_include_index(sample_df):