+++ docs/source/release_notes.rst
@@ -15,6 +15,7 @@ Release Notes
         * Update ``DataTable.get_mutual_information`` to work with Dask input (:pr:`300`)
         * Modify ``to_pandas`` function to return DataFrame with correct index (:pr:`281`)
         * Rename ``DataColumn.to_pandas`` method to ``DataColumn.to_series`` (:pr:`311`)
+        * Rename ``DataTable.to_pandas`` method to ``DataTable.to_dataframe`` (:pr:`319`)
     * Documentation Changes
         * Create a guide for using Woodwork with Dask (:pr:`304`)
         * Add conda install instructions (:pr:`305`, :pr:`309`)
@@ -28,6 +29,7 @@ Release Notes
 **Breaking Changes**
 
     * The ``DataColumn.to_pandas`` method was renamed to ``DataColumn.to_series``. 
+    * The ``DataTable.to_pandas`` method was renamed to ``DataTable.to_dataframe``. 
 
 **v0.0.4** October 21, 2020
     * Enhancements